backend/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
SQLAlchemy==2.0.0
psycopg2-binary==2.9.6
requests==2.31.0
PyYAML==6.0.1
